Provide LSP server initializationOptions support
This is to add support for LSP server initializationOptions.
For me it was important to allow transmission of the licenseKey to the
intelephense PHP language server, however it is a standard field and
other servers configurations may require these too.

Some of configurations are (or may be) expected to be determined by the
editor ('LSP client') however this should require individual per-server
'wrappers' or something similar.

* Add 'initializationOptions' as LSP server configuration option.
* Check if 'initializationOptions' are not empty and if so transmit
  these to the LSP server
* Updated README.md
* Updated plugin documentation