Ignore invalid lnum and column for prop_add
Diagnostics arrive asynchronous and the document might have changed
before they reach the client.

This could cause the line number or column to be invalid when the
diagnostics arrive.

Ignoring these errors isn't a big deal as new diagnostics will be send
by the server shortly. And when there isn't any document changes, the
client will be updated ASAP.