Add ":LspShowServer initializeRequest"
Since it's now possible to add "rootSearch" which changes the rootUri
that is being send to the server I think it makes sense to be able to
easily get to this information. I imagine it will be useful when people
are reporting issues.

It might make sense to request people, that are doing bug reports, to
include the information from:

	:LspShowServer status
	:LspShowServer initializeRequest
	:LspShowServer capabilities