Take the original sorting into account when sorting textedits
Some language servers like "efm-langserver" configured with "prettier"
will provide multiple textedits for the same line and column, these are
actually sorted "correctly" but since we apply the changes in reverse we
need to also take the original sorting into account.

The range provided is something like:

    { start: { line: 10, character: 0 }, end: { line: 10, character: 0 } }

and this is the LSP servers way to provide an insert.