Prioritize diagnostics by severity when rendering
For signs the diagnostic with the lowest (error is lowest) severity, in
a given line, is selected as being the severity to render in the sign
column.

For inline and virtual text diagnostics, the same is true, but here it
only happens if multiple diagnostics are at the same position (line +
character), this mostly happens if using multiple servers.