Skip invalid servers instead of aborting in LspAddServer()
If the user calls LspAddServer() with a list of servers, where the very
first entry does not exist, but all the remaining ones do, then no
servers are added at all. This doesn't make much sense, as the other
entries might be valid entries.

Fix this by simply continuing with the next entry instead of returning
when encountering an error.