Add option to limit triggering of auto completion
I find it distracting to have auto completion for everything I type.
I'd rather have it only triggered on trigger characters, e.g. when
typing a dot to access a class member, or a colon, when using a
namespaced symbol.

This change allows to flexibly limit auto completion to any trigger
kinds.