Fix diagnostics reporting in vala-lanauge-server
This may also affect other language servers, but vala-language-server
sends textDocument/publishDiagnostics messages for the parent directory
of the current file, *after* sending diagnostics for the current file.

As the parent directory will be a match for the current buffer using
`bufnr()` (matches a file-pattern, not an exact match on buffer name),
but the diagnostics for the parent directory are empty, the diagnostics
for the current file are initially shown and then immediately cleared.

Fix by using `bufexists()` for exact match on buffer name.

Fixes #717